Added most of the features
Finished adding most of the new buttons for checkin and home page. Also added activity card to the checkinDetails component

TODO: still need to add routing to the activity card so it can be dynamic and need to add a way to recommend activity depending on the emotion that is shown on the component.