Added assets and more style adjusting
Added the assets and text for the second healthy habits page. Also found a way to pick the correct healthy habit page. Continued to adjust styles.

NOTE: Need to somehow fix the styling on filter bar for the filtered activity screen. For some reason there is a gap of space on some resolutions